feat: #64 장소 검색까지 정상작동 확인

+++ moyeo/moyeo/Model/MeetingModel.swift
@@ -10,7 +10,7 @@ import Foundation
 struct MeetingResponse {
     
     struct Meeting: Codable {
-        let meetingId: Int64
+        let meetingId: Int
     }
 }
